fix: scope not-found to (main) group to avoid duplicate Navbar in dev (Turbopack) - Added not-found.tsx — group-scoped not-found that does not render Navbar. - Removed not-found.tsx (global not-found) to avoid double rendering.
